requirements.txt
scrapy==2.11.1
celery[redis]==5.3.6
redis==5.0.4
pymongo==4.6.1
pyrogram==2.0.106
tgcrypto==1.2.5
python-dotenv==1.0.1
